developo

+++ common/include/common/packets/CastTargetCommand.h
@@ -11,7 +11,7 @@ public:
 	virtual void Read(std::vector<uint8_t>* data) override;
 	virtual void Write(std::vector<uint8_t>* data) override;
 
-	int spell_slot;
+	uint64_t spell_slot;
 	UnitId target;
 };
