+++ common/include/common/packets/LobbySlotCommand.h
@@ -8,5 +8,5 @@ public:
 	virtual void Read(std::vector<uint8_t>* data) override;
 	virtual void Write(std::vector<uint8_t>* data) override;
 
-	int slot;
+	uint8_t slot;
 };
